Rename the RHF folder to HF.
