+++ gqcp/include/QCMethod/HF/RHFDensityMatrixCalculation.hpp
@@ -15,24 +15,24 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#pragma once
+#pragma once 
 
 
 #include "Mathematical/Algorithm/IterationStep.hpp"
-#include "QCMethod/RHF/RHFSCFEnvironment.hpp"
-#include "QCMethod/RHF/RHF.hpp"
+#include "QCMethod/HF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHF.hpp"
 
 
 namespace GQCP {
 
 
 /**
- *  An iteration step that calculates the current electronic RHF energy.
+ *  An iteration step that calculates the current density matrix (expressed in the scalar/AO basis) from the current coefficient matrix.
  * 
  *  @tparam _Scalar              the scalar type used to represent the expansion coefficient/elements of the transformation matrix
  */
 template <typename _Scalar>
-class RHFElectronicEnergyCalculation : 
+class RHFDensityMatrixCalculation :
     public IterationStep<RHFSCFEnvironment<_Scalar>> {
 
 public:
@@ -47,18 +47,14 @@ public:
      */
 
     /**
-     *  Calculate the current electronic RHF energy and place it in the environment
+     *  Calculate the current RHF density matrix and place it in the environment
      * 
      *  @param environment              the environment that acts as a sort of calculation space
      */
     void execute(Environment& environment) override {
-
-        const auto& D = environment.density_matrices.back();  // the most recent density matrix
-        const ScalarSQOneElectronOperator<Scalar> F ({environment.fock_matrices.back()});  // the most recent Fock matrix
-        const auto& H_core = environment.sq_hamiltonian.core();  // the core Hamiltonian matrix
-
-        const auto E_electronic = calculateRHFElectronicEnergy(D, H_core, F);
-        environment.electronic_energies.push_back(E_electronic);
+        const auto& C = environment.coefficient_matrices.back();  // the most recent coefficient matrix
+        const auto D = calculateRHFAO1RDM(C, environment.N);
+        environment.density_matrices.push_back(D);
     }
 };
