+++ gqcp/include/QCMethod/HF/RHFErrorCalculation.hpp
@@ -19,20 +19,20 @@
 
 
 #include "Mathematical/Algorithm/IterationStep.hpp"
-#include "QCMethod/RHF/RHF.hpp"
-#include "QCMethod/RHF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHF.hpp"
 
 
 namespace GQCP {
 
 
 /**
- *  An iteration step that calculates the current Fock matrix (expressed in the scalar/AO basis) from the current density matrix.
+ *  An iteration step that calculates the error matrix from the Fock and density matrices (expressed in the scalar/AO basis).
  * 
  *  @tparam _Scalar              the scalar type used to represent the expansion coefficient/elements of the transformation matrix
  */
 template <typename _Scalar>
-class RHFFockMatrixCalculation :
+class RHFErrorCalculation :
     public IterationStep<RHFSCFEnvironment<_Scalar>> {
 
 public:
@@ -47,14 +47,20 @@ public:
      */
 
     /**
-     *  Calculate the current RHF density matrix (expressed in the scalar/AO basis) and place it in the environment
+     *  Calculate the current error vector and add it to the environment.
      * 
      *  @param environment              the environment that acts as a sort of calculation space
      */
     void execute(Environment& environment) override {
-        const auto& D = environment.density_matrices.back();  // the most recent density matrix
-        const auto F = calculateRHFAOFockMatrix(D, environment.sq_hamiltonian);
-        environment.fock_matrices.push_back(F.parameters());
+
+        // Read F, D and S from the environment
+        const auto& D = environment.density_matrices.back();
+        const auto& S = environment.S;
+        const auto& F = environment.fock_matrices.back();
+
+        // Calculate the error and write it to the environment (as a vector)
+        const auto error_matrix = calculateRHFError(F, D, S);
+        environment.error_vectors.push_back(error_matrix.pairWiseReduce());
     }
 };
