+++ gqcp/include/QCMethod/HF/RHFFockMatrixCalculation.hpp
@@ -15,25 +15,24 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#pragma once 
+#pragma once
 
 
 #include "Mathematical/Algorithm/IterationStep.hpp"
-#include "QCMethod/RHF/RHFSCFEnvironment.hpp"
-
-#include <Eigen/Dense>
+#include "QCMethod/HF/RHF.hpp"
+#include "QCMethod/HF/RHFSCFEnvironment.hpp"
 
 
 namespace GQCP {
 
 
 /**
- *  An iteration step that solves the generalized eigenvalue problem for the current scalar/AO basis Fock matrix for the coefficient matrix.
+ *  An iteration step that calculates the current Fock matrix (expressed in the scalar/AO basis) from the current density matrix.
  * 
  *  @tparam _Scalar              the scalar type used to represent the expansion coefficient/elements of the transformation matrix
  */
 template <typename _Scalar>
-class RHFFockMatrixDiagonalization :
+class RHFFockMatrixCalculation :
     public IterationStep<RHFSCFEnvironment<_Scalar>> {
 
 public:
@@ -48,21 +47,14 @@ public:
      */
 
     /**
-     *  Solve the generalized eigenvalue problem for the most recent scalar/AO Fock matrix. Add the associated coefficient matrix and orbital energies to the environment.
+     *  Calculate the current RHF density matrix (expressed in the scalar/AO basis) and place it in the environment
      * 
      *  @param environment              the environment that acts as a sort of calculation space
      */
     void execute(Environment& environment) override {
-
-        const auto& F = environment.fock_matrices.back();  // the most recent scalar/AO basis Fock matrix
-
-        using MatrixType = Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic>;
-        Eigen::GeneralizedSelfAdjointEigenSolver<MatrixType> generalized_eigensolver (F, environment.S);
-        const TransformationMatrix<Scalar>& C = generalized_eigensolver.eigenvectors();
-        const auto& orbital_energies = generalized_eigensolver.eigenvalues();
-
-        environment.coefficient_matrices.push_back(C);
-        environment.orbital_energies.push_back(orbital_energies);
+        const auto& D = environment.density_matrices.back();  // the most recent density matrix
+        const auto F = calculateRHFAOFockMatrix(D, environment.sq_hamiltonian);
+        environment.fock_matrices.push_back(F.parameters());
     }
 };
