+++ gqcp/include/QCMethod/RMP2/RMP2.hpp
@@ -20,7 +20,7 @@
 
 #include "Molecule/Molecule.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
-#include "QCMethod/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF.hpp"
 
 
 namespace GQCP {
