+++ gqcp/src/QCMethod/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(Applications)
 add_subdirectory(CI)
 add_subdirectory(Geminals)
+add_subdirectory(HF)
 add_subdirectory(OrbitalOptimization)
-add_subdirectory(RHF)
 add_subdirectory(RMP2)
