+++ gqcp/src/QCMethod/HF/RHF.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "QCMethod/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF.hpp"
 
 
 namespace GQCP {
