+++ gqcp/tests/QCMethod/Applications/AtomicDecompositionParameters_test.cpp
@@ -26,7 +26,7 @@
 #include "Processing/RDM/RDMCalculator.hpp"
 #include "QCMethod/CI/CISolver.hpp"
 #include "QCMethod/CI/HamiltonianBuilder/FCI.hpp"
-#include "QCMethod/RHF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHFSCFSolver.hpp"
 #include "Utilities/units.hpp"
