+++ gqcp/tests/QCMethod/Geminals/AP1roG_test.cpp
@@ -25,7 +25,7 @@
 #include "Processing/Properties/expectation_values.hpp"
 #include "QCMethod/Geminals/AP1roGPSESolver.hpp"
 #include "QCMethod/Geminals/AP1roGLagrangianOptimizer.hpp"
-#include "QCMethod/RHF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHFSCFSolver.hpp"
 
 
 /**
