+++ gqcp/tests/QCMethod/HF/RHF_test.cpp
@@ -19,7 +19,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "QCMethod/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF.hpp"
 
 
 BOOST_AUTO_TEST_CASE ( RHF_1RDM_invalid_argument ) {
